Added: Some messy code
This implements more of the protocol. Specifically, the status byte is retrieved based on the last Data stage (wherein a timestamps collection is updated with the server UUID and a timestamp) and the amount of time left to wait. The time left to wait is also calculated based on the timestamps collection ((last data + 30 minutes) - Current time).

In other news, more exceptions are caught. The Client now has a Request and the server's UUID tied to it. There is an enum to handle the status byte. Encoding Strings to ByteBufs is its own method now. Added a modifier. Added a new key to the config.